# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Appointment Management System API
 * OpenAPI spec version: 0.1.0
 */
export type HealthStatusStatus =
  (typeof HealthStatusStatus)[keyof typeof HealthStatusStatus];

export const HealthStatusStatus = {
  ok: "ok",
} as const;

export interface HealthStatus {
  status: HealthStatusStatus;
}

export interface OkResponse {
  ok: boolean;
}

export interface ErrorResponse {
  error: string;
}

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Appointment {
  id: string;
  fullName: string;
  address: string;
  phoneNumber: string;
  /** @nullable */
  email?: string | null;
  reason: string;
  requestedStaff: string;
  appointmentDate: string;
  status: AppointmentStatus;
  createdAt: string;
  updatedAt: string;
}

export interface Patient {
  phoneNumber: string;
  fullName: string;
  /** @nullable */
  email?: string | null;
  appointmentCount: number;
  /** @nullable */
  lastAppointment?: string | null;
  appointments: Appointment[];
}

export type ChatbotStateStep =
  (typeof ChatbotStateStep)[keyof typeof ChatbotStateStep];

export const ChatbotStateStep = {
  greeting: "greeting",
  ask_name: "ask_name",
  ask_address: "ask_address",
  ask_phone: "ask_phone",
  ask_email: "ask_email",
  ask_reason: "ask_reason",
  ask_staff: "ask_staff",
  ask_datetime: "ask_datetime",
  confirm: "confirm",
  done: "done",
} as const;

export type ChatbotStateData = {
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  requestedStaff?: string | null;
  /** @nullable */
  appointmentDate?: string | null;
};

/**
 * Conversation state passed back to the server on each turn.
 */
export interface ChatbotState {
  step: ChatbotStateStep;
  data: ChatbotStateData;
}

/**
 * Preferred reply language. Defaults to en.
 */
export type ChatbotRequestLang =
  (typeof ChatbotRequestLang)[keyof typeof ChatbotRequestLang];

export const ChatbotRequestLang = {
  en: "en",
  am: "am",
} as const;

export interface ChatbotRequest {
  /** The user's text input. Omit/empty to start a new conversation. */
  message?: string;
  state: ChatbotState;
  /** Preferred reply language. Defaults to en. */
  lang?: ChatbotRequestLang;
}

export type ChatbotResponseInputType =
  (typeof ChatbotResponseInputType)[keyof typeof ChatbotResponseInputType];

export const ChatbotResponseInputType = {
  text: "text",
  datetime: "datetime",
  choice: "choice",
  none: "none",
} as const;

export interface ChatbotResponse {
  /** Bot's message back to the user. */
  reply: string;
  state: ChatbotState;
  /** Suggested quick-reply options (may be empty). */
  options: string[];
  inputType?: ChatbotResponseInputType;
  /** The created appointment, present only after successful booking. */
  appointment?: Appointment | null;
  /** @nullable */
  error?: string | null;
  /**
   * True when the bot's reply is a validation rejection or correction.
   * @nullable
   */
  isError?: boolean | null;
}

export interface AvailabilityResult {
  available: boolean;
  message?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface AdminUser {
  id: string;
  username: string;
  displayName: string;
}

export interface LoginResponse {
  user: AdminUser;
}

export interface SessionResponse {
  user?: AdminUser | null;
}

export interface StatusUpdateRequest {
  status: AppointmentStatus;
}

export interface AdminStats {
  total: number;
  pending: number;
  confirmed: number;
  completed: number;
  cancelled: number;
  upcoming: number;
}

export type CheckAvailabilityParams = {
  requestedStaff: string;
  appointmentDate: string;
};

export type ListAdminAppointmentsParams = {
  staff?: string;
  /**
   * ISO date (YYYY-MM-DD) to filter to a single day
   */
  date?: string;
  status?: AppointmentStatus;
};