# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Api
  version: 0.1.0
  description: AI-Powered Appointment Management System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: chatbot
  - name: appointments
  - name: auth
  - name: admin

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /chatbot/message:
    post:
      operationId: chatbotMessage
      tags: [chatbot]
      summary: Send a message to the booking chatbot. Returns next prompt or final confirmation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatbotRequest"
      responses:
        "200":
          description: Bot response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatbotResponse"

  /appointments/check-availability:
    get:
      operationId: checkAvailability
      tags: [appointments]
      summary: Check whether a specific staff member is free at a given time
      parameters:
        - in: query
          name: requestedStaff
          required: true
          schema: { type: string }
        - in: query
          name: appointmentDate
          required: true
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResult"

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Manager login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getCurrentAdmin
      tags: [auth]
      summary: Get the currently logged in manager (or null)
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"

  /auth/logout:
    post:
      operationId: adminLogout
      tags: [auth]
      summary: Log out the current manager
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  /admin/appointments:
    get:
      operationId: listAdminAppointments
      tags: [admin]
      summary: List all appointments with optional filters
      parameters:
        - in: query
          name: staff
          required: false
          schema: { type: string }
        - in: query
          name: date
          required: false
          schema: { type: string, format: date }
          description: ISO date (YYYY-MM-DD) to filter to a single day
        - in: query
          name: status
          required: false
          schema:
            $ref: "#/components/schemas/AppointmentStatus"
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/appointments/{id}/status:
    put:
      operationId: updateAppointmentStatus
      tags: [admin]
      summary: Update an appointment's status (confirm, complete, cancel)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdateRequest"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/appointments/{id}:
    delete:
      operationId: deleteAppointment
      tags: [admin]
      summary: Delete an appointment
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Summary stats for the dashboard
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/patients:
    get:
      operationId: listPatients
      tags: [admin]
      summary: List all unique patients with their appointment history
      responses:
        "200":
          description: Patient list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]

    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    AppointmentStatus:
      type: string
      enum: [pending, confirmed, completed, cancelled]

    Appointment:
      type: object
      required:
        - id
        - fullName
        - address
        - phoneNumber
        - reason
        - requestedStaff
        - appointmentDate
        - status
        - createdAt
        - updatedAt
      properties:
        id: { type: string }
        fullName: { type: string }
        address: { type: string }
        phoneNumber: { type: string }
        email: { type: string, nullable: true }
        reason: { type: string }
        requestedStaff: { type: string }
        appointmentDate: { type: string, format: date-time }
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    Patient:
      type: object
      required: [phoneNumber, fullName, appointmentCount, appointments]
      properties:
        phoneNumber: { type: string }
        fullName: { type: string }
        email: { type: string, nullable: true }
        appointmentCount: { type: integer }
        lastAppointment: { type: string, format: date-time, nullable: true }
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"

    ChatbotState:
      type: object
      description: Conversation state passed back to the server on each turn.
      required: [step, data]
      properties:
        step:
          type: string
          enum:
            - greeting
            - ask_name
            - ask_address
            - ask_phone
            - ask_email
            - ask_reason
            - ask_staff
            - ask_datetime
            - confirm
            - done
        data:
          type: object
          properties:
            fullName: { type: string, nullable: true }
            address: { type: string, nullable: true }
            phoneNumber: { type: string, nullable: true }
            email: { type: string, nullable: true }
            reason: { type: string, nullable: true }
            requestedStaff: { type: string, nullable: true }
            appointmentDate: { type: string, format: date-time, nullable: true }

    ChatbotRequest:
      type: object
      required: [state]
      properties:
        message:
          type: string
          description: The user's text input. Omit/empty to start a new conversation.
        state:
          $ref: "#/components/schemas/ChatbotState"
        lang:
          type: string
          enum: [en, am]
          description: Preferred reply language. Defaults to en.

    ChatbotResponse:
      type: object
      required: [reply, state, options]
      properties:
        reply:
          type: string
          description: Bot's message back to the user.
        state:
          $ref: "#/components/schemas/ChatbotState"
        options:
          type: array
          items: { type: string }
          description: Suggested quick-reply options (may be empty).
        inputType:
          type: string
          enum: [text, datetime, choice, none]
        appointment:
          allOf:
            - $ref: "#/components/schemas/Appointment"
          nullable: true
          description: The created appointment, present only after successful booking.
        error:
          type: string
          nullable: true
        isError:
          type: boolean
          nullable: true
          description: True when the bot's reply is a validation rejection or correction.

    AvailabilityResult:
      type: object
      required: [available]
      properties:
        available: { type: boolean }
        message: { type: string }

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }

    AdminUser:
      type: object
      required: [id, username, displayName]
      properties:
        id: { type: string }
        username: { type: string }
        displayName: { type: string }

    LoginResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/AdminUser"

    SessionResponse:
      type: object
      properties:
        user:
          allOf:
            - $ref: "#/components/schemas/AdminUser"
          nullable: true

    StatusUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/AppointmentStatus"

    AdminStats:
      type: object
      required: [total, pending, confirmed, completed, cancelled, upcoming]
      properties:
        total: { type: integer }
        pending: { type: integer }
        confirmed: { type: integer }
        completed: { type: integer }
        cancelled: { type: integer }
        upcoming: { type: integer }