# lib/api-zod/src/generated/types/appointment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Appointment Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentStatus } from "./appointmentStatus";

export interface Appointment {
  id: string;
  fullName: string;
  address: string;
  phoneNumber: string;
  /** @nullable */
  email?: string | null;
  reason: string;
  requestedStaff: string;
  appointmentDate: Date;
  status: AppointmentStatus;
  createdAt: Date;
  updatedAt: Date;
}