# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Appointment Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { Appointment } from "./appointment";

export interface Patient {
  phoneNumber: string;
  fullName: string;
  /** @nullable */
  email?: string | null;
  appointmentCount: number;
  /** @nullable */
  lastAppointment?: Date | null;
  appointments: Appointment[];
}